pyproject.toml
[project]
name = "discord-telegram-voice-notify"
version = "0.1.0"
description = "Posts to Telegram when someone joins a Discord voice channel"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.5",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.14.13",
    "ty>=0.0.12",
]
